requirements.txt
aiohttp
aiosqlite
discord.py
python-dotenv
ollama>=0.3.0
cryptography>=41.0.0
aiofiles>=23.0.0
python-dateutil>=2.8.0